Cleanup, split a few modules up
